priority queue in own file
