+++ ogame.go
@@ -40,48 +40,6 @@ const defaultUserAgent = "" +
 // CelestialID represent either a PlanetID or a MoonID
 type CelestialID int
 
-// item ...
-type item struct {
-	canBeProcessedCh chan struct{}
-	isDoneCh         chan struct{}
-	priority         int
-	index            int // The index of the item in the heap.
-}
-
-// A priorityQueue implements heap.Interface and holds Items.
-type priorityQueue []*item
-
-func (pq priorityQueue) Len() int { return len(pq) }
-
-func (pq priorityQueue) Less(i, j int) bool {
-	// We want Pop to give us the highest, not lowest, priority so we use greater than here.
-	return pq[i].priority > pq[j].priority
-}
-
-func (pq priorityQueue) Swap(i, j int) {
-	pq[i], pq[j] = pq[j], pq[i]
-	pq[i].index = i
-	pq[j].index = j
-}
-
-// Push ...
-func (pq *priorityQueue) Push(x interface{}) {
-	n := len(*pq)
-	item := x.(*item)
-	item.index = n
-	*pq = append(*pq, item)
-}
-
-// Pop ...
-func (pq *priorityQueue) Pop() interface{} {
-	old := *pq
-	n := len(old)
-	item := old[n-1]
-	item.index = -1 // for safety
-	*pq = old[0 : n-1]
-	return item
-}
-
 // OGame is a client for ogame.org. It is safe for concurrent use by
 // multiple goroutines (thread-safe)
 type OGame struct {
